Moved project to Xcode 9 for submission, added alert before opening deep link
