+++ Shortcuts Gallery/Controller/GalleryTableViewController.swift
@@ -1,0 +1,61 @@
+//
+//  ViewController.swift
+//  Shortcuts Gallery
+//
+//  Created by Marco Capano on 09/07/2018.
+//  Copyright © 2018 Marco Capano. All rights reserved.
+//
+
+import UIKit
+
+class GalleryTableViewController: UITableViewController {
+    
+    let dataManager = GalleryDataManager()
+    let searchController = UISearchController(searchResultsController: nil)
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        configureNavBar()
+        configureTableView()
+        load()
+    }
+    
+    @objc private func showAbout() {
+        let about = AboutTableViewController()
+        navigationController?.pushViewController(about, animated: true)
+    }
+    
+    @objc private func load() {
+        Shortcut.loadLatest { (response) in
+            DispatchQueue.main.async {
+                self.dataManager.shortcuts = response.results
+                self.tableView.reloadData()
+                self.refreshControl?.endRefreshing()
+            }
+        }
+    }
+    
+    private func configureTableView() {
+        tableView.register(UINib(nibName: "GalleryTableViewCell", bundle: nil), forCellReuseIdentifier: "galleryCell")
+        tableView.dataSource = dataManager
+        tableView.delegate = dataManager
+        dataManager.viewController = self
+        tableView.separatorStyle = .none
+        tableView.contentInset.top = 30
+        
+        refreshControl = UIRefreshControl()
+        refreshControl?.addTarget(self, action: #selector(load), for: .valueChanged)
+    }
+    
+    private func configureNavBar() {
+        title = "Gallery"
+        navigationItem.largeTitleDisplayMode = .always
+        navigationController?.navigationBar.prefersLargeTitles = true
+        
+        let infoButton = UIButton(type: .infoLight)
+        infoButton.addTarget(self, action: #selector(showAbout), for: .touchUpInside)
+        navigationItem.rightBarButtonItem = UIBarButtonItem(customView: infoButton)
+    }
+}
+
