+++ Shortcuts Gallery/Model/Shortcut.swift
