+++ Shortcuts Gallery/Model/ShortcutDetail.swift
