+++ Shortcuts Gallery/Model/ShortcutResponse.swift
