+++ Shortcuts Gallery/View/Base.lproj/LaunchScreen.storyboard
