+++ Shortcuts Gallery/View/GalleryTableViewCell.swift
