+++ Shortcuts Gallery/View/GalleryTableViewCell.xib
