+++ Supporting Files/Assets.xcassets/twitter.imageset/Fill 1.png
